整理程式碼 將face.py和main.py所有處理臉部偵測的搬到class MasterDetector處理
